requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
requests==2.32.5
PyYAML==6.0.2
